wavesurfer.js: separate types from src
