+++ types/wavesurfer.js/index.d.ts
@@ -6,22 +6,6 @@
 //                 Claas Augner <https://github.com/caugner>
 // Definitions: https://github.com/DefinitelyTyped/DefinitelyTyped
 
-/// <reference path="./backend.d.ts" />
-/// <reference path="./canvas.d.ts" />
-/// <reference path="./common.d.ts" />
-/// <reference path="./plugin.d.ts" />
-/// <reference path="./observer.d.ts" />
-/// <reference path="./params.d.ts" />
-/// <reference path="./util.d.ts" />
-/// <reference path="./wavesurfer.d.ts" />
-
-import * as wavesurfer from "./wavesurfer";
+import WaveSurfer = require('./src/wavesurfer');
 
 export = WaveSurfer;
-
-declare class WaveSurfer extends wavesurfer.WaveSurfer {}
-
-declare namespace WaveSurfer {
-    type PluginParams = wavesurfer.PluginParams;
-    type PluginDefinition = wavesurfer.PluginDefinition;
-}
