+++ types/wavesurfer.js/src/mediaelement-webaudio.d.ts
@@ -1,0 +1,13 @@
+import { WaveSurferParams } from "../types/params";
+import MediaElement = require("./mediaelement");
+
+export = MediaElementWebAudio;
+
+declare class MediaElementWebAudio extends MediaElement {
+    constructor(params: WaveSurferParams);
+
+    createMediaElementSource(mediaElement: HTMLMediaElement): void;
+    destroy(): void;
+    init(): void;
+    play(start: number, end: number): Promise<void>;
+}
