+++ types/wavesurfer.js/src/peakcache.d.ts
@@ -1,0 +1,9 @@
+export = PeakCache;
+
+declare class PeakCache {
+    constructor();
+
+    addRangeToPeakCache(length: number, start: number, end: number): number[][];
+    clearPeakCache(): void;
+    getCacheRanges(): number[][];
+}
