--- types/wavesurfer.js/wavesurfer.d.ts
@@ -1,100 +1,0 @@
-import CursorPlugin = require("./src/plugin/cursor");
-import ElanPlugin = require("./src/plugin/elan");
-import MarkersPlugin = require("./src/plugin/markers");
-import MediaSessionPlugin = require("./src/plugin/mediasession");
-import MicrophonePlugin = require("./src/plugin/microphone");
-import MinimapPlugin = require("./src/plugin/minimap");
-import PlayheadPlugin = require("./src/plugin/playhead");
-import RegionsPlugin = require("./src/plugin/regions");
-import SpectogramPlugin = require("./src/plugin/spectogram");
-import TimelinePlugin = require("./src/plugin/timeline");
-
-declare module './wavesurfer' {
-    class WaveSurfer extends Observer {
-        constructor(params: WaveSurferParams);
-
-        static VERSION: string;
-        static util: WaveSurfer['util'];
-        static create(params: WaveSurferParams): WaveSurfer;
-
-        backend: WaveSurferBackend;
-        util: WaveSurferUtil;
-
-        [x: string]: any;
-
-        // pluginName -> WaveSurferPlugin
-
-        /** Only available in WaveSurfer instances with CursorPlugin. */
-        cursor: CursorPlugin;
-        /** Only available in WaveSurfer instances with ElanPlugin. */
-        elan: ElanPlugin;
-        /** Only available in WaveSurfer instances with MarkersPlugin. */
-        markers: MarkersPlugin;
-        /** Only available in WaveSurfer instances with MediaSessionPlugin. */
-        mediasession: MediaSessionPlugin;
-        /** Only available in WaveSurfer instances with MicrophonePlugin. */
-        microphone: MicrophonePlugin;
-        /** Only available in WaveSurfer instances with MinimapPlugin. */
-        minimap: MinimapPlugin;
-        /** Only available in WaveSurfer instances with PlayheadPlugin. */
-        playhead: PlayheadPlugin;
-        /** Only available in WaveSurfer instances with RegionsPlugin. */
-        regions: RegionsPlugin;
-        /** Only available in WaveSurfer instances with SpectogramPlugin. */
-        spectogram: SpectogramPlugin;
-        /** Only available in WaveSurfer instances with TimelinePlugin. */
-        timeline: TimelinePlugin;
-
-        // propertyNameAddedByPlugin -> any
-
-        addPlugin(plugin: PluginDefinition): WaveSurfer;
-        cancelAjax(): void;
-        destroy(): void;
-        destroyPlugin(name: string): WaveSurfer;
-        empty(): void;
-        exportImage(format?: string, quality?: number, type?: "dataURL" | "blob"): string | string[] | Promise<Blob[]>;
-        exportPCM(length?: number, accuracy?: number, noWindow?: boolean, start?: number, end?: number): Promise<Peaks>;
-        getActivePlugins(): { [pluginName: string]: boolean };
-        getBackgroundColor(): string;
-        getCurrentTime(): number;
-        getCursorColor(): string;
-        getDuration(): number;
-        getFilters(): AudioNode[];
-        getHeight(): number;
-        getMute(): boolean;
-        getPlaybackRate(): number;
-        getProgressColor(): string;
-        getVolume(): number;
-        getWaveColor(): CanvasGradient | string;
-        init(): WaveSurfer;
-        initPlugin(name: string): WaveSurfer;
-        isPlaying(): boolean;
-        isReady: boolean;
-        load(url: string | HTMLMediaElement, peaks?: Peaks, preload?: string, duration?: number): void;
-        loadBlob(url: Blob | File): void;
-        pause(): Promise<void> | undefined;
-        play(start?: number, end?: number): Promise<void> | undefined;
-        playPause(): Promise<void> | undefined;
-        registerPlugins(plugins: PluginDefinition[]): WaveSurfer;
-        seekAndCenter(progress: number): void;
-        seekTo(progress: number): void;
-        setBackgroundColor(color: string): void;
-        setCurrentTime(seconds: number): void;
-        setCursorColor(color: string): void;
-        setHeight(height: number): void;
-        setMute(mute: boolean): void;
-        setPlaybackRate(rate: number): void;
-        setPlayEnd(position: number): void;
-        setSinkId(deviceId: string): Promise<any>;
-        setVolume(newVolume: number): void;
-        setWaveColor(color: string | CanvasGradient): void;
-        skip(offset: number): void;
-        skipBackward(seconds?: number): void;
-        skipForward(seconds?: number): void;
-        stop(): void;
-        toggleInteraction(): void;
-        toggleMute(): void;
-        toggleScroll(): void;
-        zoom(pxPerSec?: number): void;
-    }
-}
